p2p: Split MsgID into its own file as a public method
