+++ beacon-chain/p2p/BUILD.bazel
@@ -19,6 +19,7 @@ go_library(
         "interfaces.go",
         "iterator.go",
         "log.go",
+        "message_id.go",
         "monitoring.go",
         "options.go",
         "pubsub.go",
@@ -113,6 +114,7 @@ go_test(
         "fork_test.go",
         "gossip_scoring_params_test.go",
         "gossip_topic_mappings_test.go",
+        "message_id_test.go",
         "options_test.go",
         "parameter_test.go",
         "pubsub_filter_test.go",
