+++ beacon-chain/p2p/message_id_test.go
@@ -1,0 +1,66 @@
+package p2p_test
+
+import (
+	"fmt"
+	"testing"
+	"time"
+
+	"github.com/golang/snappy"
+	pubsubpb "github.com/libp2p/go-libp2p-pubsub/pb"
+	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
+	"github.com/prysmaticlabs/prysm/beacon-chain/p2p"
+	"github.com/prysmaticlabs/prysm/shared/bytesutil"
+	"github.com/prysmaticlabs/prysm/shared/hashutil"
+	"github.com/prysmaticlabs/prysm/shared/p2putils"
+	"github.com/prysmaticlabs/prysm/shared/params"
+	"github.com/prysmaticlabs/prysm/shared/testutil/assert"
+)
+
+func TestMsgID_HashesCorrectly(t *testing.T) {
+	genesisValidatorsRoot := bytesutil.PadTo([]byte{'A'}, 32)
+	d, err := p2putils.CreateForkDigest(time.Now(), genesisValidatorsRoot)
+	assert.NoError(t, err)
+	tpc := fmt.Sprintf(p2p.BlockSubnetTopicFormat, d)
+	invalidSnappy := [32]byte{'J', 'U', 'N', 'K'}
+	pMsg := &pubsubpb.Message{Data: invalidSnappy[:], Topic: &tpc}
+	hashedData := hashutil.Hash(append(params.BeaconNetworkConfig().MessageDomainInvalidSnappy[:], pMsg.Data...))
+	msgID := string(hashedData[:20])
+	assert.Equal(t, msgID, p2p.MsgID(genesisValidatorsRoot, pMsg), "Got incorrect msg id")
+
+	validObj := [32]byte{'v', 'a', 'l', 'i', 'd'}
+	enc := snappy.Encode(nil, validObj[:])
+	nMsg := &pubsubpb.Message{Data: enc, Topic: &tpc}
+	hashedData = hashutil.Hash(append(params.BeaconNetworkConfig().MessageDomainValidSnappy[:], validObj[:]...))
+	msgID = string(hashedData[:20])
+	assert.Equal(t, msgID, p2p.MsgID(genesisValidatorsRoot, nMsg), "Got incorrect msg id")
+}
+
+func TestMessageIDFunction_HashesCorrectlyAltair(t *testing.T) {
+	genesisValidatorsRoot := bytesutil.PadTo([]byte{'A'}, 32)
+	d, err := helpers.ComputeForkDigest(params.BeaconConfig().AltairForkVersion, genesisValidatorsRoot)
+	assert.NoError(t, err)
+	tpc := fmt.Sprintf(p2p.BlockSubnetTopicFormat, d)
+	topicLen := uint64(len(tpc))
+	topicLenBytes := bytesutil.Uint64ToBytesLittleEndian(topicLen)
+	invalidSnappy := [32]byte{'J', 'U', 'N', 'K'}
+	pMsg := &pubsubpb.Message{Data: invalidSnappy[:], Topic: &tpc}
+	// Create object to hash
+	combinedObj := append(params.BeaconNetworkConfig().MessageDomainInvalidSnappy[:], topicLenBytes...)
+	combinedObj = append(combinedObj, tpc...)
+	combinedObj = append(combinedObj, pMsg.Data...)
+	hashedData := hashutil.Hash(combinedObj)
+	msgID := string(hashedData[:20])
+	assert.Equal(t, msgID, p2p.MsgID(genesisValidatorsRoot, pMsg), "Got incorrect msg id")
+
+	validObj := [32]byte{'v', 'a', 'l', 'i', 'd'}
+	enc := snappy.Encode(nil, validObj[:])
+	nMsg := &pubsubpb.Message{Data: enc, Topic: &tpc}
+	// Create object to hash
+	combinedObj = append(params.BeaconNetworkConfig().MessageDomainValidSnappy[:], topicLenBytes...)
+	combinedObj = append(combinedObj, tpc...)
+	combinedObj = append(combinedObj, validObj[:]...)
+	hashedData = hashutil.Hash(combinedObj)
+	msgID = string(hashedData[:20])
+	assert.Equal(t, msgID, p2p.MsgID(genesisValidatorsRoot, nMsg), "Got incorrect msg id")
+}
+
